add docstrings to Vector3D class
